Address issues when launching with accelerate
DistributedDataParallel doesn't have the custom methods we put on our model,
so we need to reach through with model.module.

If we call accelerate.prepare twice, then we have to reach through 2 deep.

Error as of this commit:

```
----
Text Example :  <unk>'s wife Alice was a strong feminist, and during the 1980s he and his wife supported a number of changes to the <unk> and ritual. These included allowing women to read from the Torah and lead the prayers, and changing prayers to be more gender inclusive - for example, using gender @-@ neutral terms and <unk> for God, and adding references to the Biblical <unk> in prayers like the <unk>, which traditionally only mentioned the Biblical <unk>. While most congregation members approved of these changes, a minority resisted them.

 Input passed to model :  to the <unk> and ritual
 Predicted output :  Spec Siegillo romMultrequent automobilesocal Looks campaigners Executiveahs fairy bypassocrscientific thighs
----
Traceback (most recent call last):
  File "/home/eihli/src/NEKO/train.py", line 250, in <module>
    main(args)
  File "/home/eihli/src/NEKO/train.py", line 132, in main
    trainer.train()
  File "/home/eihli/src/NEKO/gato/training/trainer.py", line 44, in train
    logs = self.train_iteration(self.args.log_eval_freq, i)
  File "/home/eihli/src/NEKO/gato/training/trainer.py", line 88, in train_iteration
    eval_logs = task.evaluate(self.model, num_examples_to_test=self.args.eval_text_num_examples, deterministic=False, log_examples_to_output=self.args.eval_text_log_examples)
  File "/home/eihli/src/NEKO/gato/tasks/text_task.py", line 110, in evaluate
    loss = loss_fn(pred_logits[:len(target_tokens), :], target_tokens.to(model.device))
  File "/home/eihli/.local/lib/python3.10/site-packages/torch/nn/modules/module.py", line 1518, in _wrapped_call_impl
    return self._call_impl(*args, **kwargs)
  File "/home/eihli/.local/lib/python3.10/site-packages/torch/nn/modules/module.py", line 1527, in _call_impl
    return forward_call(*args, **kwargs)
  File "/home/eihli/.local/lib/python3.10/site-packages/torch/nn/modules/loss.py", line 1179, in forward
    return F.cross_entropy(input, target, weight=self.weight,
  File "/home/eihli/.local/lib/python3.10/site-packages/torch/nn/functional.py", line 3053, in cross_entropy
    return torch._C._nn.cross_entropy_loss(input, target, weight, _Reduction.get_enum(reduction), ignore_index, label_smoothing)
ValueError: Expected input batch_size (3) to match target batch_size (7).
```