Move argparsing to its own file
Just separation of concerns. This change supports
training in an environment that doesn't have arguments
to parse. For example: inside a Jupyter environment. There
might be a few additional changes to further support training
in an environment that doesn't have command line args. But this
felt like a good start.